Add recommended metric/span attributes
The exporter specification
[recommends](https://github.com/newrelic/newrelic-exporter-specs/blob/master/Guidelines.md#common-patterns)
setting both the `instrumentation.provider` and `collector.name`
attributes to identify the origin of telemetry. This instruments the
exporter to annotate these attributes appropriately.

This does not annotate the `instrumentation.name` attribute. This value
is something that would need to be defined by the author instrumenting
code with OpenCensus.